db migration to neon

+++ src/app/api/deleteWishlistItem/route.ts
@@ -2,10 +2,26 @@ import { getAuthSession } from "@/src/lib/auth/auth-options"
 import { db } from "@/src/server/db"
 import { wishlist, wishlistItem } from "@/src/server/db/schema"
 import { eq } from "drizzle-orm"
+import { Ratelimit } from "@upstash/ratelimit" 
+import { redis } from "@/src/server/upstash"
+import { headers } from "next/headers"
+
+const rateLimit = new Ratelimit({
+  redis,
+  limiter: Ratelimit.slidingWindow(3, "30 s"),
+  analytics: true,
+})
 
 export async function PUT(request: any) {
   try {
     const session = await getAuthSession()
+    const ip = headers().get("x-forwarded-for")
+    const {
+      remaining,
+      limit,
+      success: limitReached,
+    } = await rateLimit.limit(ip!)
+    console.log("Rate Limit Stats:", remaining, limit, limitReached)
 
     if (!session) {
       return new Response("Unauthrised request, please login.", { status: 401 })
@@ -13,13 +29,19 @@ export async function PUT(request: any) {
 
     const body = await request.json()
     const { listingId } = body
-    console.log("listingId", listingId)
 
-    await db.delete(wishlistItem).where(eq(wishlistItem.adId, listingId))
+    if (!limitReached) {
+      return new Response("API request limit reached", { status: 429 })
+    } else {
+      await db.delete(wishlistItem).where(eq(wishlistItem.adId, listingId))
 
-    return new Response(`Successfully deleted ${listingId} from the wishlist`, {
-      status: 200,
-    })
+      return new Response(
+        `Successfully deleted ${listingId} from the wishlist`,
+        {
+          status: 200,
+        }
+      )
+    }
   } catch (error) {
     console.error("Server error: Error deleteing listing to wishlist:", error)
     return new Response("Could not delete listing to wishlist.", {
